chore: rename main.go -> depviz.go (to prepare arrival of cobra)
