+++ util.go
@@ -1,0 +1,36 @@
+package main
+
+import (
+	"fmt"
+	"strings"
+)
+
+func wrap(text string, lineWidth int) string {
+	words := strings.Fields(strings.TrimSpace(text))
+	if len(words) == 0 {
+		return text
+	}
+	wrapped := words[0]
+	spaceLeft := lineWidth - len(wrapped)
+	for _, word := range words[1:] {
+		if len(word)+1 > spaceLeft {
+			wrapped += "\n" + word
+			spaceLeft = lineWidth - len(word)
+		} else {
+			wrapped += " " + word
+			spaceLeft -= 1 + len(word)
+		}
+	}
+
+	return wrapped
+}
+
+func escape(input string) string {
+	return fmt.Sprintf("%q", input)
+}
+
+func panicIfErr(err error) {
+	if err != nil {
+		panic(err)
+	}
+}
